change Config aspect from the old format (function) to the new format (class with static members)
