+++ scopes/harmony/config/index.ts
@@ -1,4 +1,3 @@
-export { Config } from './config';
-export type { ConfigMain } from './config.main.runtime';
+export { ConfigMain } from './config.main.runtime';
 export { ConfigAspect, ConfigRuntime } from './config.aspect';
 export { ComponentScopeDirMap, getWorkspaceConfigTemplateParsed, stringifyWorkspaceConfig } from './workspace-config';
