--- scopes/harmony/use/index.ts
@@ -1,5 +1,0 @@
-import { UseAspect } from './use.aspect';
-
-export type { UseMain } from './use.main.runtime';
-export default UseAspect;
-export { UseAspect };
