--- scopes/harmony/use/use.aspect.ts
@@ -1,5 +1,0 @@
-import { Aspect } from '@teambit/harmony';
-
-export const UseAspect = Aspect.create({
-  id: 'teambit.bit/use',
-});
